Reintroduce clicking command
- Tweak the click command to use a similar api to the other
commands
- Choose a random index to pick one of the two children in the
component and click on it
- Use random strings as initial values for the input
- Run tests in band: since all tests are sharing state (the DOM)
sometimes it would happen that the `cleanup` call would tore the DOM
while some othre test was using it. The concerning thing is that it
didn't make the test fail but there were errors showing; this was the
easiest way to prevent those from happening.